+++ changelog.md
@@ -1,3 +1,6 @@
 
+**<span style="color:#56adda">0.0.2</span>**
+- Add comchap and comcut to plugin
+
 **<span style="color:#56adda">0.0.1</span>**
 - Initial version